requirements.txt
deep_sort_realtime==1.3.2
deep_translator==1.11.4
easyocr==1.7.1
Flask==2.2.2
moviepy==1.0.3
numpy>=1.23.0, <2.0.0
opencv_python==4.6.0.66
opencv_python_headless==4.10.0.84
Pillow==10.2.0
qdrant_client==1.11.2
Requests==2.32.3
retry==0.9.2
scenedetect==0.6.4
torch==2.4.1
torchvision==0.19.1
transformers==4.44.2
ultralytics==8.2.87
vietocr==0.3.13
